[Refactor/erd-sync] ERD에 맞게 타입 싱크 (#37)
* ♻️ Refactor : Sync codes with Database

* ♻️ Refactor : Update test code and Resolve generic type warning in test

* ♻️ Refactor : Add Profile and CreationTimeAudit

* ♻️ Refactor : Adjust types in test code

* ♻️ Refactor : Adjust totalDistance name in RunningRecord and Remove updated_at in Profile

* ♻️ Refactor : Fix profile and user relationships and Apply to test code

* ♻️ Refactor : Fix to separate the roles of fixtures and repositories.

+++ src/main/java/com/run_us/server/domains/running/repository/PersonalRecordRepository.java
@@ -17,7 +17,7 @@ public interface PersonalRecordRepository extends JpaRepository<PersonalRecord, 
             "p.createdAt) " +
             "FROM PersonalRecord p " +
             "WHERE p.userId = :userId AND p.runningId = :runningId")
-    Optional<PersonalRecordQueryResult> findByUserIdAndRunningId(Long userId, Long runningId);
+    Optional<PersonalRecordQueryResult> findByUserIdAndRunningId(Integer userId, Integer runningId);
 
     @Query("SELECT new com.run_us.server.domains.running.service.model.PersonalRecordQueryResult(" +
             "p.runningDistanceInMeters, " +
@@ -26,5 +26,5 @@ public interface PersonalRecordRepository extends JpaRepository<PersonalRecord, 
             "p.createdAt) " +
             "FROM PersonalRecord p " +
             "WHERE p.userId = :userId")
-    List<PersonalRecordQueryResult> findAllByUserId(Long userId);
+    List<PersonalRecordQueryResult> findAllByUserId(Integer userId);
 }
